+++ R/dtplyr-package.R
@@ -1,8 +1,9 @@
 #' @import rlang
 #' @importFrom data.table data.table as.data.table .SD copy is.data.table
+#' @importFrom lifecycle deprecated
 "_PACKAGE"
 
 #' @export
 .datatable.aware <- TRUE
 
-globalVariables(".N")
+globalVariables(c(".N", "desc"))