+++ R/step-call.R
@@ -37,8 +37,13 @@ tail.dtplyr_step <- function(x, n = 6L, ...) {
 #' @importFrom dplyr rename
 #' @export
 rename.dtplyr_step <- function(.data, ...) {
-  vars <- tidyselect::vars_rename(.data$vars, ...)
-  new_vars <- names(vars)
+  sim_data <- simulate_vars(.data)
+  locs <- tidyselect::eval_rename(expr(c(...)), sim_data)
+
+  new_vars <- .data$vars
+  new_vars[locs] <- names(locs)
+
+  vars <- set_names(.data$vars[locs], names(locs))
   vars <- vars[vars != names(vars)]
 
   if (length(vars) == 0) {
@@ -56,6 +61,19 @@ rename.dtplyr_step <- function(.data, ...) {
   step_group(out, groups)
 }
 
+#' @importFrom dplyr rename_with
+#' @importFrom tidyselect everything
+#' @export
+rename_with.dtplyr_step <- function(.data, .fn, .cols = everything(), ...) {
+  .fn <- as_function(.fn)
+  sim_data <- simulate_vars(.data)
+  locs <- tidyselect::eval_select(enquo(.cols), sim_data)
+
+  to_change <- .data$vars[locs]
+  names(to_change) <- .fn(to_change, ...)
+
+  rename(.data, !!!to_change)
+}
 
 #' @importFrom dplyr distinct
 #' @export