+++ tests/testthat/test-step-group.R
@@ -7,13 +7,16 @@ test_that("grouping and ungrouping adjust groups field", {
   expect_equal(dt %>% group_by(x) %>% ungroup() %>% .$groups, character())
 })
 
-
 test_that("can add groups if requested", {
   dt <- lazy_dt(data.frame(x = 1:3, y = 1:3), "DT")
   expect_equal(
-    dt %>% group_by(x) %>% group_by(y, add = TRUE) %>% .$groups,
+    dt %>% group_by(x) %>% group_by(y, .add = TRUE) %>% .$groups,
     c("x", "y")
   )
+
+  expect_snapshot({
+    . <- dt %>% group_by(x) %>% group_by(y, add = TRUE)
+  })
 })
 
 test_that("grouping can compute new variables if needed", {